fix ProduceStrategyConfig and AgentLoopConfig
