+++ xtuner/v1/rl/base/agent_loop.py
@@ -1,6 +1,6 @@
 import asyncio
 from abc import ABC, abstractmethod
-from typing import Callable, Type
+from typing import Callable
 
 from pydantic import BaseModel, ConfigDict
 
@@ -8,6 +8,25 @@ from xtuner.v1.data_proto import RolloutState, SampleParams, Status
 from xtuner.v1.ray.judger import NativeJudger, NativeJudgerRouter
 from xtuner.v1.ray.rollout import RolloutController
 from xtuner.v1.utils.processing_utils import load_processor, load_tokenizer
+
+
+class AgentLoopConfig(ABC, BaseModel):
+    model_config = ConfigDict(extra="allow", arbitrary_types_allowed=True)
+    hf_checkpoint: str
+    sample_params: SampleParams
+
+    @abstractmethod
+    def build(self, rollout_controller, judger=None) -> "AgentLoop": ...
+
+
+class SingleTurnAgentLoopConfig(AgentLoopConfig):
+    def build(self, rollout_controller, judger=None) -> "SingleTurnAgentLoop":
+        return SingleTurnAgentLoop(
+            rollout_ctl=rollout_controller,
+            hf_checkpoint=self.hf_checkpoint,
+            sample_params=self.sample_params,
+            judger=judger,
+        )
 
 
 class AgentLoop(ABC):
@@ -59,18 +78,3 @@ class SingleTurnAgentLoop(AgentLoop):
             return rollout_state
         rollout_state = await self.judge_sample(rollout_state)
         return rollout_state
-
-
-class AgentLoopConfig(BaseModel):
-    model_config = ConfigDict(extra="allow", arbitrary_types_allowed=True)
-    type: Type[AgentLoop] = SingleTurnAgentLoop
-    hf_checkpoint: str
-    sample_params: SampleParams
-
-    def build(self, rollout_controller, judger=None) -> AgentLoop:
-        return self.type(
-            rollout_ctl=rollout_controller,
-            sample_params=self.sample_params,
-            hf_checkpoint=self.hf_checkpoint,
-            judger=judger,
-        )
