+++ xtuner/v1/rl/base/agent_loop_manager.py
@@ -7,6 +7,27 @@ from xtuner.v1.rl.base.replay_buffer import ReplayBuffer
 from .agent_loop import AgentLoop
 
 
+class AgentLoopManagerConfig(BaseModel):
+    model_config = ConfigDict(extra="forbid", arbitrary_types_allowed=True)
+
+    task_name: str
+
+    def build(
+        self,
+        agent_loop: AgentLoop,
+        produce_strategy: ProduceStrategy,
+        sampler: Sampler,
+        replay_buffer: ReplayBuffer,
+    ) -> "AgentLoopManager":
+        return AgentLoopManager(
+            agent_loop=agent_loop,
+            produce_strategy=produce_strategy,
+            sampler=sampler,
+            replay_buffer=replay_buffer,
+            task_name=self.task_name,
+        )
+
+
 class AgentLoopManager:
     def __init__(
         self,
@@ -37,24 +58,3 @@ class AgentLoopManager:
     # async def disaggregate_get_batch(self, task_name: str, batch_size: int):
     #     # 从不同的 replay_buffer 中采样，然后训练
     #     return self._replay_buffer.get(batch_size, task_name, Status.COMPLETED)
-
-
-class AgentLoopManagerConfig(BaseModel):
-    model_config = ConfigDict(extra="forbid", arbitrary_types_allowed=True)
-
-    task_name: str
-
-    def build(
-        self,
-        agent_loop: AgentLoop,
-        produce_strategy: ProduceStrategy,
-        sampler: Sampler,
-        replay_buffer: ReplayBuffer,
-    ) -> AgentLoopManager:
-        return AgentLoopManager(
-            agent_loop=agent_loop,
-            produce_strategy=produce_strategy,
-            sampler=sampler,
-            replay_buffer=replay_buffer,
-            task_name=self.task_name,
-        )
