+++ xtuner/v1/rl/base/producer.py
@@ -1,74 +1,69 @@
 import asyncio
 from abc import ABC, abstractmethod
-from typing import Callable, Iterator, Literal, Optional, Type
+from typing import Protocol, runtime_checkable
 
 from pydantic import BaseModel, ConfigDict
 
-from transformers import AutoTokenizer, PreTrainedTokenizer, PreTrainedTokenizerFast
 from xtuner.v1.data_proto.rl_data import RolloutState, Status
-from xtuner.v1.datasets.config import DataloaderConfig
-from xtuner.v1.datasets.dataloader import Dataloader
 
 from .agent_loop import AgentLoop
 from .replay_buffer import ReplayBuffer
-
-
-class _DatasetSampler:
-    def __init__(self, dataloader: Dataloader, prompt_repeat_k: int):
-        self.dataloader = dataloader
-        self.dataloader_iter: Optional[Iterator] = None
-        self.cur_epoch = 0
-        self.prompt_repeat_k = prompt_repeat_k
-
-    def sample_from_dataloader(self) -> list[RolloutState]:
-        if self.dataloader_iter is None:
-            self.dataloader_iter = iter(self.dataloader)
-        assert self.dataloader_iter is not None
-        try:
-            data = next(self.dataloader_iter)[0]
-        except StopIteration:
-            self.cur_epoch += 1
-            self.dataloader.set_epoch(self.cur_epoch)
-            self.dataloader_iter = iter(self.dataloader)
-            data = next(self.dataloader_iter)[0]
-        group_data = [data] * self.prompt_repeat_k
-        return group_data
-
-
-class Sampler(_DatasetSampler):
-    def __init__(
-        self,
-        dataloader: Dataloader,
-        prompt_repeat_k: int,
-        replay_buffer: ReplayBuffer,
-    ):
-        super().__init__(dataloader, prompt_repeat_k)
-        self.replay_buffer = replay_buffer
-
-    async def sample(self, task_name: str) -> list[RolloutState]:
-        buffer_data = await self.replay_buffer.get(1, task_name=task_name, group_status=Status.ABORTED)
-        if len(buffer_data) == 0:
-            return self.sample_from_dataloader()
-        else:
-            return buffer_data[0]
+from .sampler import Sampler
 
 
 def default_is_valid_sample_fn(samples: list[RolloutState]) -> bool:
     return all(sample.status == Status.COMPLETED for sample in samples)
 
 
-def default_should_continue_fn(completed_count: int, batch_size: int) -> bool:
+def default_should_continue_fn(completed_count: int, batch_size: int, **kwargs) -> bool:
     return completed_count < batch_size
+
+
+@runtime_checkable
+class IsValidSampleFn(Protocol):
+    def __call__(self, samples: list[RolloutState]) -> bool: ...
+
+
+@runtime_checkable
+class ShouldContinueFn(Protocol):
+    def __call__(self, completed_count: int, batch_size: int, **kwargs) -> bool: ...
+
+
+class ProducerConfig(ABC, BaseModel):
+    model_config = ConfigDict(extra="forbid", arbitrary_types_allowed=True)
+    is_valid_sample_fn: IsValidSampleFn = default_is_valid_sample_fn
+    should_continue_fn: ShouldContinueFn = default_should_continue_fn
+
+    @abstractmethod
+    def build(self) -> "ProduceStrategy": ...
+
+
+class SyncProduceStrategyConfig(ProducerConfig):
+    def build(self) -> "SyncProduceStrategy":
+        return SyncProduceStrategy(
+            is_valid_sample_fn=self.is_valid_sample_fn, should_continue_fn=self.should_continue_fn
+        )
+
+
+class OverProduceStrategyConfig(ProducerConfig):
+    staleness_threshold: float = 0.0
+
+    def build(self) -> "OverProduceStrategy":
+        return OverProduceStrategy(
+            staleness_threshold=self.staleness_threshold,
+            is_valid_sample_fn=self.is_valid_sample_fn,
+            should_continue_fn=self.should_continue_fn,
+        )
 
 
 class ProduceStrategy(ABC):
     def __init__(
         self,
-        is_valid_sample_fn: Optional[Callable[[list[RolloutState]], bool]] = None,
-        should_continue_fn: Optional[Callable[[int, int], bool]] = None,
+        is_valid_sample_fn: IsValidSampleFn,
+        should_continue_fn: ShouldContinueFn,
     ):
-        self.is_valid_sample_fn = is_valid_sample_fn or default_is_valid_sample_fn
-        self.should_continue_fn = should_continue_fn or default_should_continue_fn
+        self.is_valid_sample_fn = is_valid_sample_fn
+        self.should_continue_fn = should_continue_fn
 
     @abstractmethod
     async def produce_batch(
@@ -125,9 +120,9 @@ class SyncProduceStrategy(ProduceStrategy):
 class OverProduceStrategy(ProduceStrategy):
     def __init__(
         self,
-        staleness_threshold: float = 0.0,
-        is_valid_sample_fn: Optional[Callable[[list[RolloutState]], bool]] = None,
-        should_continue_fn: Optional[Callable[[int, int], bool]] = None,
+        staleness_threshold: float,
+        is_valid_sample_fn: IsValidSampleFn,
+        should_continue_fn: ShouldContinueFn,
     ):
         super().__init__(is_valid_sample_fn, should_continue_fn)
         self.staleness_threshold = staleness_threshold
@@ -183,46 +178,3 @@ class OverProduceStrategy(ProduceStrategy):
                     await agent_loop.pause()
                     await asyncio.sleep(1)
         print("All worker tasks have completed after pausing env controller.")
-
-
-class SamplerConfig(BaseModel):
-    model_config = ConfigDict(extra="forbid", arbitrary_types_allowed=True)
-    dataloader_cfg: DataloaderConfig
-    prompt_repeat_k: int = 1
-
-    def build(
-        self, tokenizer: PreTrainedTokenizer | PreTrainedTokenizerFast | str, replay_buffer: ReplayBuffer
-    ) -> Sampler:
-        if isinstance(tokenizer, str):
-            tokenizer_obj = AutoTokenizer.from_pretrained(tokenizer, trust_remote_code=True)
-        else:
-            tokenizer_obj = tokenizer
-        dataloader = self.dataloader_cfg.build(
-            tokenizer=tokenizer_obj, dp_mesh=None, global_batch_size=1, micro_batch_size=1, seed=1
-        )
-        return Sampler(dataloader=dataloader, prompt_repeat_k=self.prompt_repeat_k, replay_buffer=replay_buffer)
-
-
-class ProducerConfig(BaseModel):
-    model_config = ConfigDict(extra="forbid", arbitrary_types_allowed=True)
-    # TODO: 要考虑怎么方便用户使用自定义的 ProduceStrategy, 并且怎么使用用户需要的参数
-    type: Type[ProduceStrategy] | Literal["sync", "over"] = "sync"
-    staleness_threshold: float = 0.0  # only used for over_produce strategy
-    is_valid_sample_fn: Optional[Callable[[list[RolloutState]], bool]] = None
-    should_continue_fn: Optional[Callable[[int, int], bool]] = None
-
-    def build(self) -> "ProduceStrategy":
-        if self.type == "sync":
-            if self.staleness_threshold != 0.0:
-                print("Warning: staleness_threshold is ignored in sync produce strategy.")
-            return SyncProduceStrategy(
-                is_valid_sample_fn=self.is_valid_sample_fn, should_continue_fn=self.should_continue_fn
-            )
-        elif self.type == "over":
-            return OverProduceStrategy(
-                staleness_threshold=self.staleness_threshold,
-                is_valid_sample_fn=self.is_valid_sample_fn,
-                should_continue_fn=self.should_continue_fn,
-            )
-        else:
-            return self.type(is_valid_sample_fn=self.is_valid_sample_fn, should_continue_fn=self.should_continue_fn)
